Lag tilbake BearerTokenClientInterceptor for prosjekter som ikke har behov for å skille på grantType.
